+++ lib/packages/core/src/internal/index.ts
@@ -1,3 +1,5 @@
 export * from './signAndBroadcastTransaction';
+export * from './constructAttachment';
+export * from './createParametersFromAttachment';